requirements.txt
boto3==1.34.149
GitPython==3.1.43
langchain==0.2.11
langchain-aws @ git+https://github.com/3coins/langchain-aws.git@fix-anthropic-tool-call#subdirectory=libs/aws
langchain_community==0.2.10
langchain-core==0.2.26
langchain-text-splitters==0.2.2
langchain_openai==0.1.20
langgraph==0.1.14
litellm==1.41.6
prompt_toolkit==3.0.47
pytest==8.3.2
python-dotenv==1.0.1
rich==13.7.1
